requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
cffi==1.15.1
channels==3.0.4
channels-redis==4.0.0
constantly==15.1.0
cryptography==40.0.1
daphne==3.0.2
Django==4.2
django-ckeditor==6.5.1
django-js-asset==2.0.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.1.1
redis==4.5.4
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
zope.interface==6.0
